first commit
--- changes: ---

new project files -- refactoring in the future is a must

--- thoughts: ---

Starting on a calendar project. I'm not sure that I am doing things the right way,
but It seems like an interesting challenge, making a calendar daemon that should
be cross platform. At the current moment I will only be working on Pluto, until
it is at a complete enough state that I can start work on the front end for it.